Cleaned up IModel
